Pick some low-hanging warning fruit by changing several integer variables to size_t or apr_size_t, depending on the context of usage.  This removes a number of "implicit conversion shortens 64-bit value into a 32-bit value" warnings.
* subversion/svn/util.c
  (svn_cl__get_log_message),
* subversion/libsvn_client/merge.c
  (fix_deleted_subtree_ranges, populate_remaining_ranges,
   record_mergeinfo_for_dir_merge),
* subversion/libsvn_client/locking_commands.c
  (organize_lock_targets),
* subversion/libsvn_client/commit_util.c
  (svn_client__condense_commit_items),
* subversion/libsvn_client/add.c
  (trim_string, auto_props_enumerator):
  Update variable types.

* subversion/svn/main.c
  (main):
  Cast value to (int).


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@879075 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_client/add.c
@@ -78,7 +78,7 @@ static void
 trim_string(char **pstr)
 {
   char *str = *pstr;
-  int i;
+  size_t i;
 
   while (apr_isspace(*str))
     str++;
@@ -160,7 +160,7 @@ auto_props_enumerator(const char *name,
 
   for (i = 0; i < props->nelts; i++)
     {
-      int len;
+      size_t len;
       const char *this_value;
       char *property = APR_ARRAY_IDX(props, i, char *);
       char *equal_sign = strchr(property, '=');
